# ui/valset-standardizer-frontend/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Java 外部估值表标准化工程，基于 Spring Boot 和 Quartz",
    "version": "v1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:30066"
    }
  ],
...
  "paths": {
...
    "/api/transfer-sources/template-name": {
      "get": {
        "summary": "查询来源对应表单模板名。",
        "deprecated": false,
        "tags": [
          "TransferSourceController"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SingleResultString"
                }
              }
            }
          }
        },
        "operationId": "getTemplateName_1",
        "parameters": [
          {
            "schema": {
              "format": "enum",
              "type": "string",
              "enum": [
                "LOCAL_DIR",
                "EMAIL",
                "S3",
                "SFTP"
              ]
            },
            "in": "query",
            "description": "来源类型<br/>(See: 文件来源类型。)",
            "required": true,
            "example": "LOCAL_DIR",
            "name": "sourceType"
          }
        ]
      }
    },
...
    "/api/transfer-targets": {
      "post": {
        "summary": "创建投递目标。",
        "deprecated": false,
        "tags": [
          "TransferTargetController"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransferTargetUpsertCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SingleResultTransferTargetMutationResponse"
                }
              }
            }
          }
        },
        "operationId": "createTarget",
        "parameters": []
      },
      "get": {
        "summary": "查询投递目标列表。",
        "deprecated": false,
        "tags": [
          "TransferTargetController"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MultiResultTransferTargetViewDTO"
                }
              }
            }
          }
        },
        "operationId": "listTargets",
        "parameters": [
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "query",
            "description": "目标类型",
            "required": false,
            "name": "targetType"
          },
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "query",
            "description": "目标编码",
            "required": false,
            "name": "targetCode"
          },
          {
            "schema": {
              "format": "boolean",
              "type": "boolean"
            },
            "in": "query",
            "description": "是否启用",
            "required": false,
            "example": true,
            "name": "enabled"
          },
          {
            "schema": {
              "format": "int32",
              "type": "integer"
            },
            "in": "query",
            "description": "查询上限",
            "required": false,
            "example": 0,
            "name": "limit"
          }
        ]
      }
    },
    "/api/transfer-targets/{targetId}": {
      "get": {
        "summary": "查询投递目标详情。",
        "deprecated": false,
        "tags": [
          "TransferTargetController"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SingleResultTransferTargetViewDTO"
                }
              }
            }
          }
        },
        "operationId": "getTarget",
        "parameters": [
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "path",
            "description": "目标主键",
            "required": true,
            "name": "targetId"
          }
        ]
      },
      "delete": {
        "summary": "删除投递目标。",
        "deprecated": false,
        "tags": [
          "TransferTargetController"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SingleResultTransferTargetMutationResponse"
                }
              }
            }
          }
        },
        "operationId": "deleteTarget",
        "parameters": [
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "path",
            "description": "目标主键",
            "required": true,
            "name": "targetId"
          }
        ]
      },
      "put": {
        "summary": "更新投递目标。",
        "deprecated": false,
        "tags": [
          "TransferTargetController"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransferTargetUpsertCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SingleResultTransferTargetMutationResponse"
                }
              }
            }
          }
        },
        "operationId": "updateTarget",
        "parameters": [
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "path",
            "description": "目标主键",
            "required": true,
            "name": "targetId"
          }
        ]
      }
    },
    "/api/transfer-targets/template-name": {
      "get": {
        "summary": "查询投递目标对应表单模板名。",
        "deprecated": false,
        "tags": [
          "TransferTargetController"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SingleResultString"
                }
              }
            }
          }
        },
        "operationId": "getTemplateName_2",
        "parameters": [
          {
            "schema": {
              "format": "enum",
              "type": "string",
              "enum": [
                "EMAIL",
                "S3",
                "SFTP",
                "FILESYS"
              ]
            },
            "in": "query",
            "description": "目标类型<br/>(See: 文件投递目标类型。)",
            "required": true,
            "example": "EMAIL",
            "name": "targetType"
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "SingleResultTransferTargetViewDTO": {
        "type": "object",
        "properties": {
          "success": {
            "format": "boolean",
            "description": "状态",
            "type": "boolean",
            "example": true
          },
          "dataType": {
            "format": "string",
            "description": "返回数据格式",
            "type": "string"
          },
          "code": {
            "description": "返回状态码(object)",
            "type": "object"
          },
          "message": {
            "format": "string",
            "description": "返回消息",
            "type": "string"
          },
          "tips": {
            "format": "string",
            "description": "返回提示消息",
            "type": "string"
          },
          "data": {
            "description": "返回的单条数据(object)",
            "$ref": "#/components/schemas/TransferTargetViewDTO"
          }
        }
      },
...
      "SingleResultTransferTargetMutationResponse": {
        "type": "object",
        "properties": {
          "success": {
            "format": "boolean",
            "description": "状态",
            "type": "boolean",
            "example": true
          },
          "dataType": {
            "format": "string",
            "description": "返回数据格式",
            "type": "string"
          },
          "code": {
            "description": "返回状态码(object)",
            "type": "object"
          },
          "message": {
            "format": "string",
            "description": "返回消息",
            "type": "string"
          },
          "tips": {
            "format": "string",
            "description": "返回提示消息",
            "type": "string"
          },
          "data": {
            "description": "返回的单条数据(object)",
            "$ref": "#/components/schemas/TransferTargetMutationResponse"
          }
        }
      },
...
      "SingleResultString": {
        "type": "object",
        "properties": {
          "success": {
            "format": "boolean",
            "description": "状态",
            "type": "boolean",
            "example": true
          },
          "dataType": {
            "format": "string",
            "description": "返回数据格式",
            "type": "string"
          },
          "code": {
            "description": "返回状态码(object)",
            "type": "object"
          },
          "message": {
            "format": "string",
            "description": "返回消息",
            "type": "string"
          },
          "tips": {
            "format": "string",
            "description": "返回提示消息",
            "type": "string"
          },
          "data": {
            "format": "string",
            "description": "返回的单条数据",
            "type": "string"
          }
        }
      },
...
      "TransferTargetUpsertCommand": {
        "type": "object",
        "properties": {
          "targetId": {
            "format": "string",
            "description": "投递目标主键。",
            "type": "string"
          },
          "targetCode": {
            "format": "string",
            "description": "投递目标编码。",
            "type": "string"
          },
          "targetName": {
            "format": "string",
            "description": "投递目标名称。",
            "type": "string"
          },
          "targetType": {
            "format": "string",
            "description": "投递目标类型。",
            "type": "string"
          },
          "enabled": {
            "format": "boolean",
            "description": "是否启用。",
            "type": "boolean",
            "example": true
          },
          "targetPathTemplate": {
            "format": "string",
            "description": "投递路径模板。",
            "type": "string"
          },
          "connectionConfig": {
            "description": "连接配置。(object)",
            "$ref": "#/components/schemas/Object"
          },
          "targetMeta": {
            "description": "目标扩展信息。(object)",
            "$ref": "#/components/schemas/Object"
          }
        },
        "required": [
          "targetCode",
          "targetName",
          "targetType"
        ]
      },
...
      "Object": {
        "type": "object",
        "properties": {
          "mapKey": {
            "description": "A map key.(object)",
            "$ref": "#/components/schemas/Object"
          }
        }
      },
...
      "MultiResultTransferTargetViewDTO": {
        "type": "object",
        "properties": {
          "success": {
            "format": "boolean",
            "description": "状态",
            "type": "boolean",
            "example": true
          },
          "dataType": {
            "format": "string",
            "description": "返回数据格式",
            "type": "string"
          },
          "code": {
            "description": "返回状态码(object)",
            "type": "object"
          },
          "message": {
            "format": "string",
            "description": "返回消息",
            "type": "string"
          },
          "tips": {
            "format": "string",
            "description": "返回提示消息",
            "type": "string"
          },
          "data": {
            "description": "返回数据",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TransferTargetViewDTO"
            }
          }
        }
      },
...
      "TransferTargetViewDTO": {
        "type": "object",
        "properties": {
          "targetId": {
            "format": "string",
            "description": "投递目标主键。",
            "type": "string"
          },
          "targetCode": {
            "format": "string",
            "description": "投递目标编码。",
            "type": "string"
          },
          "targetName": {
            "format": "string",
            "description": "投递目标名称。",
            "type": "string"
          },
          "targetType": {
            "format": "string",
            "description": "投递目标类型。",
            "type": "string"
          },
          "formTemplateName": {
            "format": "string",
            "description": "当前使用的目标表单模板名。",
            "type": "string"
          },
          "enabled": {
            "format": "boolean",
            "description": "是否启用。",
            "type": "boolean",
            "example": true
          },
          "targetPathTemplate": {
            "format": "string",
            "description": "投递路径模板。",
            "type": "string"
          },
          "connectionConfig": {
            "description": "连接配置。(object)",
            "$ref": "#/components/schemas/Object"
          },
          "targetMeta": {
            "description": "目标扩展信息。(object)",
            "$ref": "#/components/schemas/Object"
          },
          "createdAt": {
            "format": "string",
            "description": "创建时间。",
            "type": "string",
            "example": "yyyy-MM-dd HH:mm:ss"
          },
          "updatedAt": {
            "format": "string",
            "description": "修改时间。",
            "type": "string",
            "example": "yyyy-MM-dd HH:mm:ss"
          }
        }
      },
...
      "TransferTargetMutationResponse": {
        "type": "object",
        "properties": {
          "operation": {
            "format": "string",
            "description": "操作类型。",
            "type": "string"
          },
          "message": {
            "format": "string",
            "description": "提示消息。",
            "type": "string"
          },
          "formTemplateName": {
            "format": "string",
            "description": "关联的表单模板名。",
            "type": "string"
          },
          "target": {
            "description": "投递目标详情。(object)",
            "$ref": "#/components/schemas/TransferTargetViewDTO"
          }
        }
      },
...
    }
  }
}